requirements.txt
aiohttp==3.7.4.post0
discord==1.7.3
asyncio~=3.4.3
requests==2.7.0
pycoingecko
python-telegram-bot
asyncio